+++ src/Widgets/Buttons/AdminButton.php
@@ -7,7 +7,9 @@ namespace Hirtz\Skeleton\Widgets\Buttons;
 use Hirtz\Skeleton\Assets\AdminAssetBundle;
 use Hirtz\Skeleton\Html\A;
 use Hirtz\Skeleton\Html\Span;
+use Hirtz\Skeleton\Html\Svg;
 use Hirtz\Skeleton\Widgets\Widget;
+use Override;
 use Stringable;
 use Yii;
 
@@ -17,7 +19,7 @@ class AdminButton extends Widget
     public string $buttonColorActive = '#FA7D7E';
     public string $buttonPosition = 'right';
 
-    public string $icon = '<svg viewBox="0 0 64 64"><path d="M25,26c0-0.6,0.4-1,1-1h6c0.6,0,1-0.4,1-1s-0.4-1-1-1h-6c-1.7,0-3,1.3-3,3v12c0,1.7,1.3,3,3,3h12c1.7,0,3-1.3,3-3v-6c0-0.6-0.4-1-1-1s-1,0.4-1,1v6c0,0.6-0.4,1-1,1H26c-0.6,0-1-0.4-1-1V26z"/><path d="M37.9,21.7c0.9-0.9,2.5-0.9,3.4,0l1,1c0.9,0.9,0.9,2.5,0,3.4l-8.1,8.1c-0.2,0.2-0.4,0.3-0.6,0.3L30.1,35c-0.3,0-0.6-0.1-0.8-0.3c-0.2-0.2-0.3-0.5-0.3-0.8l0.5-3.5c0-0.2,0.1-0.4,0.3-0.6L37.9,21.7z"/></svg>';
+    public string|Stringable $icon = '';
     public string $iconColor = '#fff';
     public string $iconColorActive = '#000';
 
@@ -31,6 +33,23 @@ class AdminButton extends Widget
      * @var bool whether to toggle the button opacity on hover, if `false` the button will always be visible
      */
     public bool $toggleButtonOpacity = true;
+
+    #[Override]
+    protected function configure(): void
+    {
+        $this->icon = $this->icon ?: $this->getDefaultIcon();
+        parent::configure();
+    }
+
+    protected function getDefaultIcon(): Svg
+    {
+        return Svg::make()
+            ->viewBox('0 0 64 64')
+            ->content(
+                '<path d="M25,26c0-0.6,0.4-1,1-1h6c0.6,0,1-0.4,1-1s-0.4-1-1-1h-6c-1.7,0-3,1.3-3,3v12c0,1.7,1.3,3,3,3h12c1.7,0,3-1.3,3-3v-6c0-0.6-0.4-1-1-1s-1,0.4-1,1v6c0,0.6-0.4,1-1,1H26c-0.6,0-1-0.4-1-1V26z"/>',
+                '<path d="M37.9,21.7c0.9-0.9,2.5-0.9,3.4,0l1,1c0.9,0.9,0.9,2.5,0,3.4l-8.1,8.1c-0.2,0.2-0.4,0.3-0.6,0.3L30.1,35c-0.3,0-0.6-0.1-0.8-0.3c-0.2-0.2-0.3-0.5-0.3-0.8l0.5-3.5c0-0.2,0.1-0.4,0.3-0.6L37.9,21.7z"/>',
+            );
+    }
 
     protected function renderContent(): string|Stringable
     {